Refactoring code 95% have to test

+++ building/module/commission/views.py
@@ -3,40 +3,53 @@ from django.shortcuts import render_to_response
 from django.http import HttpResponseRedirect
 from django.core.urlresolvers import reverse
 from django.contrib import messages
+from django.conf import settings
 
 from building.models import Building
-from building.module.commission.models import BuildingCommissionHeroThing, \
-                                              BuildingCommission
+from building.module.commission.models import BuildingCommissionHeroThing
 from thing.models import Thing
+from hero.models import HeroThing
 
 from building.module.commission.forms import PutForm
 
-from hero.heromanipulation import hero_init, update_capacity
-from building import buildingmanipulation
+from hero.manipulation import hero_init, in_given_location, HeroM
+from building.manipulation import BuildingM
 
-PLUGIN = 'commission'
+MODULE = 'commission'
 
 @hero_init
 def index(request, slug, 
           template_name='building/module/commission/index.html'):
     hero = request.hero
-    building = Building.objects.get(slug=slug)
-    
+    try:
+        building = Building.objects.get(slug=slug)
+    except Building.DoesNotExist:
+        return HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
+
+    buildingm = BuildingM(building, hero)
+
+    if not buildingm.is_near_building(slug):
+        return HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
+
+    buildingm.add_to_location(slug)
+
     variables = RequestContext(request, {'hero': hero,
                                          'building': building})
-    
-    buildingmanipulation.add_building_to_location(hero, building, slug)
     
     return render_to_response(template_name, variables)
 
 @hero_init
-def view(request, slug, type, template_name='building/module/commission/view.html'):
+@in_given_location
+def view(request, slug, type,
+         template_name='building/module/commission/view.html'):
     hero = request.hero
-    building = Building.objects.get(slug=slug)
+    try:
+        building = Building.objects.get(slug=slug)
+    except Building.DoesNotExist:
+        return HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
 
-    type_num = [i[0] for i in Thing.TYPES if i[1].lower() == type][0]
-    commissionherothings = BuildingCommissionHeroThing.objects.filter(
-                                            herothing__thing__type=type_num). \
+    commissionherothings = building.buildingcommissionherothing_set. \
+        filter(herothing__thing__type=eval('Thing.TYPE_' + type.upper())). \
                                                 exclude(herothing__hero=hero)
 
     variables = RequestContext(request, {'hero': hero,
@@ -46,38 +59,55 @@ def view(request, slug, type, template_name='building/module/commission/view.htm
     return render_to_response(template_name, variables)
 
 @hero_init
-def buy(request, slug, id):
+@in_given_location
+def buy(request, slug, commissionherothing_id):
     hero = request.hero
-    building = Building.objects.get(slug=slug)
+    try:
+        building = Building.objects.get(slug=slug)
+        commissionherothing = building.buildingcommissionherothing_set. \
+                                                get(id=commissionherothing_id)
+    except (Building.DoesNotExist, BuildingCommissionHeroThing.DoesNotExist):
+        return HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
     
-    commissionherothing = BuildingCommissionHeroThing.objects.get(id=id)
-    percent = BuildingCommission.objects.get(building=building).percent
     price = commissionherothing.price
-    
-    hero.money -= price
-    hero.save()
-    
-    commissionherothing.herothing.hero.money += price - \
-                                            round((price * (percent / 100)))  
-    commissionherothing.herothing.hero.save()
-    
-    commissionherothing.herothing.away = False
-    commissionherothing.herothing.hero = hero
-    commissionherothing.herothing.save()
-    
-    commissionherothing.delete()
-    
-    update_capacity(hero)
-#
-    messages.add_message(request, messages.SUCCESS, 'You buy thing.')
-    return HttpResponseRedirect(reverse('commission', args=[slug]))
 
-#Put
+    if hero.money < price:
+#
+        messages.add_message(request, messages.ERROR,
+                             'You have not enough money.')
+    else:
+
+        percent = building.buildingcommission_set.get().percent
+
+        hero.money -= price
+        hero.save()
+
+        commissionherothing.herothing.hero.money += price - \
+                                            round((price * (percent / 100)))
+        commissionherothing.herothing.hero.save()
+
+        commissionherothing.herothing.away = False
+        commissionherothing.herothing.hero = hero
+        commissionherothing.herothing.save()
+
+        commissionherothing.delete()
+
+        HeroM(hero).update_capacity()
+#
+        messages.add_message(request, messages.SUCCESS, 'You buy thing.')
+        
+    return HttpResponseRedirect(request.META.get('HTTP_REFERER'))
+
+# Put
 @hero_init
+@in_given_location
 def put(request, slug, template_name='building/module/commission/put.html'):
     hero = request.hero
-    building = Building.objects.get(slug=slug)
-    
+    try:
+        building = Building.objects.get(slug=slug)
+    except Building.DoesNotExist:
+        HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
+        
     herothings = hero.herothing_set.filter(dressed=False, away=False)
     
     variables = RequestContext(request, {'hero': hero,
@@ -87,29 +117,32 @@ def put(request, slug, template_name='building/module/commission/put.html'):
     return render_to_response(template_name, variables)
 
 @hero_init
-def put_select(request, slug, id,
-           template_name='building/module/commission/put_select.html'):
+@in_given_location
+def put_select(request, slug, herothing_id,
+               template_name='building/module/commission/put_select.html'):
     hero = request.hero
-    building = Building.objects.get(slug=slug)
-    
-    herothing = hero.herothing_set.get(id=id)
-    
+    try:
+        building = Building.objects.get(slug=slug)
+        herothing = hero.herothing_set.get(id=herothing_id, dressed=False,
+                                           away=False)
+    except (Building.DoesNotExist, HeroThing.DoesNotExist):
+        return HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
+
     if request.method == 'POST':
         form = PutForm(request.POST)
         if form.is_valid():
             herothing.away = True
             herothing.save()
             
-            commissionherothing = BuildingCommissionHeroThing(
-                                                            building=building,
-                                                        herothing=herothing,
+            commissionherothing = building.buildingcommissionherothing_set. \
+                                                    create(herothing=herothing,
                                             price=form.cleaned_data['price'])
             commissionherothing.save()
             
-            update_capacity(hero)
+            HeroM(hero).update_capacity()
 #
             messages.add_message(request, messages.SUCCESS, 'You put thing.')
-            return HttpResponseRedirect(reverse('commission', args=[slug]))
+            return HttpResponseRedirect(reverse('commission_put', args=[slug]))
     else:
         form = PutForm()
         
@@ -119,15 +152,19 @@ def put_select(request, slug, id,
                                          'form': form})
     
     return render_to_response(template_name, variables)
-#End
+# End put
 
-#Take
+# Take
 @hero_init
+@in_given_location
 def take(request, slug, template_name='building/module/commission/take.html'):
     hero = request.hero
-    building = Building.objects.get(slug=slug)
+    try:
+        building = Building.objects.get(slug=slug)
+    except Building.DoesNotExist:
+        return HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
     
-    commissionherothings = BuildingCommissionHeroThing.objects. \
+    commissionherothings = building.buildingcommissionherothing_set. \
                                                 filter(herothing__hero=hero)
     
     variables = RequestContext(request, {'hero': hero,
@@ -137,16 +174,22 @@ def take(request, slug, template_name='building/module/commission/take.html'):
     return render_to_response(template_name, variables)
 
 @hero_init
-def take_select(request, slug, id):
+@in_given_location
+def take_select(request, slug, commissionherothing_id):
     hero = request.hero
+    try:
+        building = Building.objects.get(slug=slug)
+        commissionherothing = building.buildingcommissionherothing_set. \
+                                                get(id=commissionherothing_id)
+    except (Building.DoesNotExist, BuildingCommissionHeroThing.DoesNotExist):
+        return HttpResponseRedirect(reverse(settings.URL_REVERSE_404))
     
-    commissionherothing = BuildingCommissionHeroThing.objects.get(id=id)
     commissionherothing.herothing.away = False
     commissionherothing.herothing.save()
     commissionherothing.delete()
     
-    update_capacity(hero)
+    HeroM(hero).update_capacity()
 #
     messages.add_message(request, messages.SUCCESS, 'You take thing.')
-    return HttpResponseRedirect(reverse('commission', args=[slug]))
-#End
+    return HttpResponseRedirect(reverse('commission_take', args=[slug]))
+# End take
